requirements.txt
fastapi
uvicorn
pydantic
pandas
requests
groq
streamlit
python-dotenv